+++ src/scenes/CyberScene.ts
@@ -1,6 +1,7 @@
 import Phaser from 'phaser';
 import { SaveManager } from '../engine/SaveManager';
 import { VFXManager } from '../engine/VFXManager';
+import { InputManager } from '../engine/InputManager';
 
 const TILE = 24;
 
@@ -126,6 +127,8 @@ export default class CyberScene extends Phaser.Scene {
       this.scene.pause();
       this.scene.launch('PauseScene', { scene: this.scene.key });
     });
+    InputManager.setLegacyGamepadKeyboardBridge(true);
+    this.events.once('shutdown', () => InputManager.setLegacyGamepadKeyboardBridge(false));
   }
 
   spawnGhost(x: number, y: number, color: number) {